+++ deltafi-ui/src/components/errors/AllPanel.vue
@@ -334,7 +334,7 @@ onMounted(() => {
 
 const getPersistedParams = async () => {
   let state = useStorage("errors-page-session-storage", {}, sessionStorage, { serializer: StorageSerializers.object });
-  perPage.value = state.value.perPage || 10;
+  perPage.value = state.value.perPage || 20;
 };
 
 const setPersistedParams = () => {