+++ deltafi-ui/src/components/errors/SummaryByMessagePanel.vue
@@ -233,7 +233,7 @@ const onPage = async (event) => {
 
 const getPersistedParams = async () => {
   let state = useStorage("errors-page-session-storage", {}, sessionStorage, { serializer: StorageSerializers.object });
-  perPage.value = state.value.perPage || 10;
+  perPage.value = state.value.perPage || 20;
 };
 
 const setPersistedParams = () => {