+++ deltafi-ui/CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Changed
 - Added Events page (#315)
 - Added Notifications badge (#311)
+- Updated nginx runtime image to 1.23.3
 
 ## [0.101.1]