+++ deltafi-ui/CHANGELOG.md
@@ -7,9 +7,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## [0.101.3]
 
 ### Changed
+
 - Added Events page (#315)
 - Added Notifications badge (#311)
 - Updated nginx runtime image to 1.23.3
+- Updated default page size to 20 (#319)
 
 ## [0.101.1]